+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## Unreleased
+
+- Updated `addPageIndex()` function signature for `PdfBookmarkTrait` trait.
+
 ## 2.0.2 - 2024-08-12
 
 - Added `PdfBookmarkTrait` trait.