+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Updated name space of `PdfVersionTest` class.
 - Update PHP version in `composer.json`.
 - Added `PdfVersion` enumeration.
 - Updated `addPageIndex()` function signature for `PdfBookmarkTrait` trait.